STRIPE/STORE BACK/STORE-BACKEND/rate_limit_reset_cf/requirements.txt
google-cloud-firestore==2.13.1
functions-framework==3.5.0
Flask==2.3.0
